Added reconcile state to output file
Added a decision tree which checks the reconciled state of each split and appends the status before the split. If the split is reconciled (y), then the ledger file shows it is cleared (*). If the split is cleared (c), then ledger file shows that it is "in progress" (!). Otherwise, there is nothing added to the split. The asterisk was removed from before the transaction description to prevent conflict.

Examples:
---------------

Transaction with all splits reconciled:

2015-01-01 Transaction description
    * Account 1: Subaccount 1          0.00 USD
    * Account 2: Subaccount 2          0.00 USD

Transaction with one reconciled split and one cleared split:

2015-01-01 Transaction description
    * Account 1: Subaccount 1          0.00 USD
    ! Account 2: Subaccount 2          0.00 USD

Transaction with no reconciled state on any split:

2015-01-01 Transaction description
    Account 1: Subaccount 1          0.00 USD
    Account 2: Subaccount 2          0.00 USD